Created payment related utils file for the customCard
